// frontend/src/types/task.tsx
export type TaskStatus =
  | "todo"
  | "in_progress"
  | "completed";

export type TaskPriority =
  | "low"
  | "medium"
  | "high";


/*
 * Task returned by Django API
 */
export interface Task {
  id: number;

  title: string;

  description: string;

  status: TaskStatus;

  priority: TaskPriority;

  due_date: string | null;

  created_at: string;

  updated_at: string;
}


/*
 * Data sent when creating/updating
 * a task.
 *
 * id, created_at and updated_at
 * are generated by Django, so
 * frontend does not send them.
 */
export interface TaskPayload {
  title: string;

  description: string;

  status: TaskStatus;

  priority: TaskPriority;

  due_date: string | null;
}


/*
 * Data used by TaskForm
 *
 * This is the same shape as
 * TaskPayload.
 */
export type TaskFormData =
  TaskPayload;